rename folder keyring-entries -> wallets
