+++ packages/iov-keycontrol/src/index.ts
@@ -1,5 +1,4 @@
 export { HdPaths } from "./hdpaths";
-export { Ed25519HdWallet, Ed25519Wallet, Secp256k1HdWallet } from "./keyring-entries";
 export {
   Keyring,
   LocalIdentity,
@@ -11,3 +10,4 @@ export {
   WalletSerializationString,
 } from "./keyring";
 export { UserProfile, WalletInfo } from "./userprofile";
+export { Ed25519HdWallet, Ed25519Wallet, Secp256k1HdWallet } from "./wallets";
