+++ packages/iov-keycontrol/src/wallets/ed25519hdwallet.ts
