+++ packages/iov-keycontrol/src/wallets/ed25519wallet.spec.ts
