+++ packages/iov-keycontrol/src/wallets/ed25519wallet.ts
