+++ packages/iov-keycontrol/types/wallets/ed25519hdwallet.d.ts
