+++ packages/iov-keycontrol/types/wallets/ed25519wallet.d.ts
