+++ packages/iov-keycontrol/types/wallets/slip10wallet.d.ts
